+++ .travis.yml
@@ -1,10 +1,10 @@
 language: generic
 
 env:
-  - BUILD_TYPE=RELEASE;PYTHON=2.7
-  - BUILD_TYPE=RELEASE;PYTHON=3.5
-  - BUILD_TYPE=DEBUG;PYTHON=2.7
-  - BUILD_TYPE=DEBUG;PYTHON=3.5
+  - BUILD_TYPE=RELEASE; PYTHON=2.7;
+  - BUILD_TYPE=RELEASE; PYTHON=3.5;
+  - BUILD_TYPE=DEBUG; PYTHON=2.7;
+  - BUILD_TYPE=DEBUG; PYTHON=3.5;
 
 os:
   - linux
@@ -26,7 +26,7 @@ addons:
       packages:
         - valgrind
 
-before_script:
+install:
   # Valgrind is experimental(ish) on MacOS with false positives on among others printf
   #- if [[ "$TRAVIS_OS_NAME" == "osx" && "$BUILD_TYPE" == "DEBUG" ]]; then
   #    brew update;
@@ -51,11 +51,15 @@ before_script:
 
   - conda install numpy cmake
 
+before_script:
+  - enabled="-DBUILD_PYTHON=OFF -DBUILD_MEX=OFF"
+  - if [[ -n "${PYTHON+1}" ]]; then enabled="$enabled -DBUILD_PYTHON=ON"; fi
   - cmake --version
+
+script:
   - mkdir build
-  - cd build
-  - cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE ..
+  - pushd build
+  - cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE $enabled  ..
   - make
   - export LD_LIBRARY_PATH=$PWD
-
-script: make && ctest --output-on-failure
+  - ctest --output-on-failure
