+++ appveyor.yml
@@ -14,15 +14,34 @@ platform:
     - x64
 
 environment:
+    OFF: -DBUILD_PYTHON=OFF -DBUILD_MEX=OFF
     matrix:
         - PYTHON: C:\Python27
+          LANG: -DBUILD_PYTHON=ON
         - PYTHON: C:\Python35
+          LANG: -DBUILD_PYTHON=ON
+
+matrix:
+    fast_finish: true
+
+init:
+    - INSTALL_DIR: C:\projects\segyio-install
+
+install:
+  - IF DEFINED PYTHON (IF "%platform%" == "x64" SET PYTHON=%PYTHON%-x64)
+  - IF DEFINED PYTHON %PYTHON%\python -m pip install --user numpy
+
+before_build:
+    - IF DEFINED PYTHON SET LANG=%LANG% -DPYTHON_EXECUTABLE=%PYTHON%\python.exe
 
 build_script:
-    - "%PYTHON%\\python -m pip install --user numpy"
     - cmake --version
+    - IF "%platform%" == "x64" set W64="-GVisual Studio 14 2015 Win64"
     - mkdir build
     - pushd build
-    - cmake C:\projects\SegyIO -DBUILD_MEX=OFF -DBUILD_PYTHON=ON -DCMAKE_BUILD_TYPE=%configuration% -DPYTHON_EXECUTABLE="%PYTHON%\\python.exe"
-    - cmake --build . --config %configuration%
-    - ctest -C %configuration% --output-on-failure -V
+    - cmake C:\projects\SegyIO %W64% %OFF% %LANG% -DCMAKE_BUILD_TYPE=%configuration% -DCMAKE_INSTALL_PREFIX=%INSTALL_DIR%
+    - cmake --build . --config %configuration% --target install
+
+test_script:
+    - set PATH=%INSTALL_DIR%\bin;%PATH%
+    - ctest -C %configuration% --output-on-failure
