+++ lib/src/segy.def
@@ -1,0 +1,49 @@
+LIBRARY segyio
+EXPORTS
+segy_open
+segy_mmap
+segy_flush
+segy_close
+segy_binheader_size
+segy_binheader
+segy_write_binheader
+segy_samples
+segy_sample_interval
+segy_format
+segy_get_field
+segy_get_bfield
+segy_set_field
+segy_set_bfield
+segy_trace_bsize
+segy_trace0
+segy_traces
+segy_sample_indexes
+segy_read_textheader
+segy_textheader_size
+segy_write_textheader
+segy_traceheader
+segy_write_traceheader
+segy_sorting
+segy_offsets
+segy_offset_indices
+segy_readtrace
+segy_writetrace
+segy_to_native
+segy_from_native
+segy_read_line
+segy_write_line
+segy_count_lines
+segy_lines_count
+segy_inline_length
+segy_crossline_length
+segy_inline_indices
+segy_crossline_indices
+segy_line_trace0
+segy_inline_stride
+segy_crossline_stride
+segy_seek
+segy_ftell
+ebcdic2ascii
+ascii2ebcdic
+ieee2ibm
+ibm2ieee
