+++ lib/test/spec.c
