+++ mex/segyutil.h
@@ -8,8 +8,32 @@
 
 #include <segyio/segy.h>
 
-#include "spec/segyspec.h"
+typedef struct {
+    char* filename;
 
+    unsigned int sample_format;
+
+    unsigned int* crossline_indexes;
+    unsigned int crossline_count;
+
+    unsigned int* inline_indexes;
+    unsigned int inline_count;
+
+    unsigned int offset_count;
+
+    double* sample_indexes;
+    unsigned int sample_count;
+
+    int trace_sorting_format;
+
+    unsigned int il_stride;
+    unsigned int xl_stride;
+    unsigned int first_trace_pos;
+    unsigned int trace_bsize;
+
+} SegySpec;
+
+int segyCreateSpec(SegySpec* spec, const char* file, unsigned int inline_field, unsigned int crossline_field, double t0, double dt);
 
 void recreateSpec(SegySpec* spec, const mxArray* mex_spec);
