+++ python/CMakeLists.txt
@@ -1,8 +1,10 @@
+project(segyio-python)
+
 if (NOT BUILD_PYTHON)
     return()
 endif()
 
-include(${CMAKE_SOURCE_DIR}/cmake/python.cmake)
+include(python)
 
 if (NOT DEFINED PYTHON_EXECUTABLE)
     message("Python interpreter not found - Python wrappers not enabled")
@@ -24,29 +26,44 @@ if (NOT PYTHONLIBS_FOUND)
     return()
 endif()
 
-include_directories(${PYTHON_INCLUDE_DIRS})
-add_library(_segyio SHARED segyio/_segyio.c $<TARGET_OBJECTS:segyio-object>)
+add_library(_segyio MODULE segyio/_segyio.c)
+target_include_directories(_segyio PRIVATE ${PYTHON_INCLUDE_DIRS})
+target_link_libraries(_segyio segyio-shared ${PYTHON_LIBRARIES})
+export(TARGETS _segyio segyio-shared FILE segyio-config.cmake)
 
-if (MSVC)
-    set(LIB_SUFFIX "pyd")
-    set(LINK_FLAGS "/DLL ${PYTHON_LIBRARY}")
-elseif (MAC_OS)
-    set(LIB_SUFFIX "so")
-    set(LINK_FLAGS "-undefined dynamic_lookup")
-else()
-    set(LIB_SUFFIX "so")
-    set(LINK_FLAGS "-Xlinker -export-dynamic")
-endif()
+add_python_package(pysegyio segyio
+                    TARGETS _segyio
+                    SOURCES segyio/__init__.py
+                            segyio/_header.py
+                            segyio/_line.py
+                            segyio/_field.py
+                            segyio/_trace.py
+                            segyio/_raw_trace.py
+                            segyio/segy.py
+                            segyio/tracefield.py
+                            segyio/binfield.py
+                            segyio/open.py
+                            segyio/tools.py
+                            segyio/create.py
+                            segyio/segysampleformat.py
+                            segyio/tracesortingformat.py)
 
-set_target_properties(_segyio PROPERTIES PREFIX "" SUFFIX ".${LIB_SUFFIX}")
-set_target_properties(_segyio PROPERTIES LINK_FLAGS ${LINK_FLAGS})
-target_link_libraries(_segyio ${SEGYIO_LIBRARIES})
+file(GLOB sgys "${CMAKE_SOURCE_DIR}/test-data/*.sgy")
+foreach (sgy ${sgys})
+    get_filename_component(fl ${sgy} NAME)
+    configure_file(${sgy} test-data/${fl} COPYONLY)
+endforeach ()
 
-install(TARGETS _segyio DESTINATION ${CMAKE_INSTALL_PREFIX}/${PYTHON_INSTALL_PREFIX}/segyio)
+configure_file(test/test_context.py test/test_context.py COPYONLY)
+add_python_test(python.segy         test/segy.py)
+add_python_test(python.h.segy       test/segyio_c.py)
+add_python_test(python.enum.segy    test/segyioenum.py)
+add_python_test(python.tools        test/tools.py)
+add_python_test(python.context      test/context.py)
 
-add_custom_command(TARGET _segyio POST_BUILD
-        COMMAND ${CMAKE_COMMAND} -E copy
-        $<TARGET_FILE:_segyio>
-        ${CMAKE_BINARY_DIR}/python/segyio/_segyio.${LIB_SUFFIX})
-
-add_subdirectory(segyio)
+add_python_example(pysegyio python.example.about         examples/about.py test-data/small.sgy INLINE_3D CROSSLINE_3D)
+add_python_example(pysegyio python.example.write         examples/write.py test-data/small.sgy)
+add_python_example(pysegyio python.example.makefile      examples/make-file.py test-data/large-file.sgy 20 1 20 1 20)
+add_python_example(pysegyio python.example.makepsfile    examples/make-ps-file.py test-data/small-prestack.sgy 10 1 5 1 4 1 3)
+add_python_example(pysegyio python.example.subcube       examples/copy-sub-cube.py test-data/small.sgy test-data/copy.sgy)
+add_python_example(pysegyio python.example.scan_min_max  examples/scan_min_max.py test-data/small.sgy)
