+++ python/examples/copy-sub-cube.py
