+++ python/examples/scan_min_max.py
