+++ python/examples/write.py
