+++ python/test/segy.py
