+++ python/test/segyio_c.py
