+++ python/test/segyioenum.py
