+++ test-data/small.sgy
